Change dealers.phone column type.

+++ db/migrate/20211211135943_create_dealers.rb
@@ -2,7 +2,7 @@ class CreateDealers < ActiveRecord::Migration[6.0]
   def change
     create_table :dealers do |t|
       t.string :name, null: false
-      t.integer :phone, null: false
+      t.bigint :phone, null: false
       t.timestamps
     end
